feat(codegen): implement function codegen and improve error handling
Refactor function code generation by moving implementation to a new
`function` module and enabling `Statement::Function` to generate code.
Fix LLVM type conversion by passing context reference correctly.
Replace error printing with panic on return build failure to avoid silent
errors. Clean up imports and remove redundant code for better maintainability.
