+++ crates/why_lib/src/codegen/expressions/num.rs
@@ -14,7 +14,7 @@ impl<'ctx> CodeGen<'ctx> for Num<ValidatedTypeInformation> {
             Num::Integer(val, info, ..) => {
                 let ValidatedTypeInformation { type_id, .. } = info;
 
-                let llvm_type = type_id.to_llvm_type(&ctx.context).into_int_type();
+                let llvm_type = type_id.to_llvm_type(ctx.context).into_int_type();
 
                 llvm_type.const_int(*val, false).into()
             }
